+++ app/controllers/character_log_entries_controller.rb
@@ -138,7 +138,7 @@ class CharacterLogEntriesController < LogEntriesController
                   :player_dm_id, :notes,
                   magic_items_attributes: [:id, :name, :rarity,
                                            :location_found, :table,
-                                           :table_result,
+                                           :table_result, :character_id,
                                            :notes, :_destroy])
   end
 end