+++ app/controllers/dm_log_entries_controller.rb
@@ -135,7 +135,7 @@ class DmLogEntriesController < LogEntriesController
                   :date_played, :character_id,
                   magic_items_attributes: [:id, :name, :rarity,
                                            :location_found, :table,
-                                           :table_result,
+                                           :table_result, :character_id,
                                            :notes, :_destroy])
   end
 end