+++ app/models/magic_item.rb
@@ -1,7 +1,7 @@
 class MagicItem < ActiveRecord::Base
   belongs_to :log_entry
   belongs_to :trade_log_entry
-  has_one :character, through: :log_entry
+  has_one :character
 
   enum rarity: [:common, :uncommon, :rare, :very_rare, :legendary]
 end