+++ db/migrate/20170330105802_add_character_id_to_magic_items.rb
@@ -1,0 +1,10 @@
+class AddCharacterIdToMagicItems < ActiveRecord::Migration
+  def change
+    add_column :magic_items, :character_id, :integer
+
+    MagicItem.all.each do |mi|
+      mi.destroy and next unless mi.log_entry
+      mi.update!(character_id: mi.log_entry.character.id)
+    end
+  end
+end