+++ db/schema.rb
@@ -11,7 +11,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 20170330053936) do
+ActiveRecord::Schema.define(version: 20170330105802) do
 
   # These are extensions that must be enabled in order to support this database
   enable_extension "plpgsql"
@@ -106,6 +106,7 @@ ActiveRecord::Schema.define(version: 20170330053936) do
     t.string  "table"
     t.string  "table_result"
     t.string  "location_found"
+    t.integer "character_id"
   end
 
   create_table "player_dms", force: :cascade do |t|