+++ CHANGELOG.md
@@ -3,6 +3,9 @@ CHANGELOG
 
 v1.4.21
 ---------------
+- Add Character selection for Magic Items in Campaign
+- Add Magic Items#index
+- Add Magic Item deletion
 - Allow multiple DM to join Campaigns
 - Allow editting of Campaigns